backend_1/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
django-filter==23.2
Pillow==10.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.5.0
drf-yasg==1.21.7
pytest==7.4.0
pytest-django==4.5.2
coverage==7.3.0
factory-boy==3.3.0
